% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gaussian.R
\docType{data}
\name{gaussian}
\alias{gaussian}
\title{A simulated gaussian data set}
\format{
A data frame with 300 rows and 8 columns.
}
\usage{
gaussian
}
\description{
A simulated gaussian data set. True parameters for this data set are given by:
\itemize{
\item Exp.1 = (0,0,1.5,0,0,0,1);
\item Exp.2 = (0,1,-1.5,0,0,2,0);
\item Gate = (1,2,0,0,-1,0,0);
\item Sigma = 1.
}
}
\keyword{datasets}
