% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.RMoE-package.R
\docType{package}
\name{GLM.RMoE-package}
\alias{GLM.RMoE}
\alias{GLM.RMoE-package}
\title{RMoE: LASSO Regularized Mixture of Experts Models}
\description{
\code{GLM.RMoE} is a package containing regularized Mixture of Experts
models using the Lasso penalty.

\code{GLM.RMoE} contains the following Regularized Mixture-of-Experts models:

\itemize{
\item GaussianRMoE: Gaussian Regularixed Mixture of Experts;
\item LogisticRMoE: Logistic Regularixed Mixture of Experts;
\item PoissonRMoE: Poisson Regularixed Mixture of Experts.
}

To learn more about \code{GLM.RMoE}, start with the vignettes:
\code{browseVignettes(package = "GLM.RMoE")}
}
\references{
Nguyen V. T., Chamroukhi F., Hoang V. H., Huynh B. T. 2026. \emph{Estimation and Feature Selection in Mixtures of Generalized Linear Experts Models.} \url{https://arxiv.org/abs/xxxx.xxxxx}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nv-thin/GLM-RMoE}
  \item Report bugs at \url{https://github.com/nv-thin/GLM-RMoE/issues}
}

}
\author{
\strong{Maintainer}: Van-Thin Nguyen \email{nvthin@hcmus.edu.vn}

Authors:
\itemize{
  \item Van-Thin Nguyen \email{nvthin@hcmus.edu.vn}
  \item Faicel Chamroukhi \email{faicel.chamroukhi@unicaen.fr} (\href{https://orcid.org/0000-0002-5894-3103}{ORCID})
  \item Van-Ha Hoang \email{hvha@hcmus.edu.vn}
  \item Bao-Tuyen Huynh \email{baotuyen.dlu@gmail.com}
}

}
