% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{The Housing data set}
\format{
A data frame with 506 rows and 15 columns.
}
\source{
\href{https://archive.ics.uci.edu/ml/machine-learning-databases/housing}{https://archive.ics.uci.edu/ml/machine-learning-databases/housing}
}
\usage{
housing
}
\description{
The Housing data set described on the website UC Irvine Machine Learning
Repository. The value MEDV/sd(MEDV) is considered as the predictor variable.
}
\keyword{datasets}
