% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ionosphere.R
\docType{data}
\name{ionosphere}
\alias{ionosphere}
\title{The Ionosphere data set}
\format{
A data frame with 351 rows and 35 columns.
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/ionosphere}{https://archive.ics.uci.edu/ml/datasets/ionosphere}
}
\usage{
ionosphere
}
\description{
The Ionosphere data set described on the website UC Irvine Machine Learning
Repository.
}
\keyword{datasets}
