% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-logistic.R
\docType{data}
\name{logistic}
\alias{logistic}
\title{A simulated logistic data set}
\format{
A data frame with 300 rows and 8 columns.
}
\usage{
logistic
}
\description{
A simulated logistic data set. True parameters for this data set are given by:
\itemize{
\item Exp.1 = (0,-1,2,0,0,1.5,0);
\item Exp.2 = (0,1,0,0,-2,0,0);
\item Gate = (1,0,0,1,0,0,-1.5).
}
}
\keyword{datasets}
