% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRMoE-class.R
\docType{class}
\name{PRMoE-class}
\alias{PRMoE-class}
\alias{PRMoE}
\title{A Reference Class which contains parameters of a PRMoE model.}
\description{
PRMoE contains all the parameters of a Poisson Regularized
Mixture-of-Experts.
}
\section{Fields}{

\describe{
\item{\code{X}}{The matrix data for the input.}

\item{\code{Y}}{Vector of the response variable.}

\item{\code{d}}{Numeric. Number of explanatory variables (including the intercept
variable).}

\item{\code{n}}{Numeric. Length of the response/output vector \code{Y}.}

\item{\code{K}}{Number of expert classes.}

\item{\code{Lambda}}{Penalty value for the expert part.}

\item{\code{Gamma}}{Penalty value for the gating network.}

\item{\code{wk}}{Parameters of the gating network. Matrix of dimension \eqn{(K - 1,
d)}, with \code{d} the number of explanatory variables (including the
intercept).}

\item{\code{betak}}{Regressions coefficients for each expert. Matrix of dimension
\eqn{(d, K)}.}

\item{\code{loglik}}{Numeric. Observed-data log-likelihood of the PRMoE model.}

\item{\code{storedloglik}}{Numeric vector. Stored values of the log-likelihood at
each EM iteration.}

\item{\code{zerocoeff}}{Matrix. Proportion of zero coefficients obtained during each
iteration of the EM. First column gives the number of zero coefficients for
\code{wk} and the second column for \code{betak}.}

\item{\code{BIC}}{Numeric. Value of BIC (Bayesian Information Criterion).}

\item{\code{Cluster}}{Numeric vector. Clustering label for each observation.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("loglik", "zerocoefficients"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"loglik" = } Value of the log-likelihood for
        each iteration.
      \item \code{"zerocoefficients" = } Proportion of zero
        coefficients for each iteration.
    }
  }
}
By default, all the above graphs are produced.}
}}

