% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-residential.R
\docType{data}
\name{residential}
\alias{residential}
\title{The Residential Building data set}
\format{
A data frame with 372 rows and 109 columns.
}
\source{
\href{http://archive.ics.uci.edu/ml/datasets/Residential+Building+Data+Set}{http://archive.ics.uci.edu/ml/datasets/Residential+Building+Data+Set}
}
\usage{
residential
}
\description{
The Residential Building data set described on the website UC Irvine Machine
Learning Repository. The V-9 variable is used as the predictor variable.
}
\keyword{datasets}
